fix(analytics): cleanups
