+++ app/components/analytics/index.ts
@@ -1,2 +1,3 @@
 export { default as AnalyticsChartCard } from "./AnalyticsChartCard";
 export { default as AnalyticsTotalsRow } from "./AnalyticsTotalsRow";
+export { default as ChartTooltip } from "./ChartTooltip";
